refactor: carry the coin type in CoinSelectionParams
CoinType is threaded through AttemptSelection and ChooseSelectionResult
as a defaulted trailing parameter. Upstream keeps rewriting exactly
these signatures (bitcoin#26661, bitcoin#27183), so every backport has
to re-resolve the Dash-only argument, and a resolution that drops the
trailing argument silently falls back to CoinType::ALL_COINS, losing
the fully-mixed selection restrictions.

Store the coin type in CoinSelectionParams instead, which is built once
per selection request and already flows through the whole pipeline. The
enum moves from coincontrol.h to coinselection.h so the field can be
declared there without pulling the much heavier coincontrol.h into the
selection module; coin control is a consumer of the selection layer, so
the include direction stays sound.
